feat: generate highlights globally
