+++ src/common/channelHighlight/definitions.ts
@@ -8,9 +8,6 @@ export const getChannelHighlightDefinitions = async ({
   con: DataSource;
 }): Promise<ChannelHighlightDefinition[]> =>
   con.getRepository(ChannelHighlightDefinition).find({
-    where: {
-      mode: Not('disabled'),
-    },
     order: {
       order: 'ASC',
       channel: 'ASC',
